requirements.txt
# Core ML and CV libraries - Updated for 2025
torch>=2.4.0
torchvision>=0.19.0
ultralytics>=8.0.196
opencv-python>=4.8.0.76

# Data processing
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.12.2
Pillow>=10.0.0

# Machine learning utilities
scikit-learn>=1.3.0
albumentations>=1.3.1

# Annotation tools
labelme>=5.2.1

# Web development
flask>=2.3.2
requests>=2.31.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0.1
